remove SpecializedAggregateOpDesc
