+++ core/amber/src/test/scala/edu/uci/ics/amber/engine/e2e/TestOperators.scala
@@ -1,11 +1,8 @@
 package edu.uci.ics.amber.engine.e2e
 
 import edu.uci.ics.amber.engine.common.workflow.PortIdentity
-import edu.uci.ics.texera.workflow.operators.aggregate.{
-  AggregationFunction,
-  AggregationOperation,
-  SpecializedAggregateOpDesc
-}
+import edu.uci.ics.texera.workflow.common.operators.aggregate.AggregateOpDesc
+import edu.uci.ics.texera.workflow.operators.aggregate.{AggregationFunction, AggregationOperation}
 import edu.uci.ics.texera.workflow.operators.hashJoin.HashJoinOpDesc
 import edu.uci.ics.texera.workflow.operators.keywordSearch.KeywordSearchOpDesc
 import edu.uci.ics.texera.workflow.operators.sink.managed.ProgressiveSinkOpDesc
@@ -84,8 +81,8 @@ object TestOperators {
       attributeToAggregate: String,
       aggFunction: AggregationFunction,
       groupByAttributes: List[String]
-  ): SpecializedAggregateOpDesc = {
-    val aggOp = new SpecializedAggregateOpDesc()
+  ): AggregateOpDesc = {
+    val aggOp = new AggregateOpDesc()
     val aggFunc = new AggregationOperation()
     aggFunc.aggFunction = aggFunction
     aggFunc.attribute = attributeToAggregate
